requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Werkzeug==3.0.3
SQLAlchemy==2.0.35
pyotp==2.9.0
cryptography==43.0.0
Pillow==10.4.0
pyzbar==0.1.9
python-dotenv==1.0.1

